Rename and group toghether modules for docker static files handling
