+++ barbeque/static_files/middleware.py
